Add Redis queues
Two implementations of a distributed queue are included.

The first is a based on an idea I have seen the Redis website where the
queue is reprented as two lists, and items are moved from one to the
other to start work and removed from the second when work is completed.
The second uses a list as a circular buffer and adds a distributed lock
to minimize redundant processing. Both provide a means of guaranteeing
at-least-once processing of each item added to the queue.